refactor: 폴더 이름 변경 및 속도 검증 리팩토링
