+++ src/classes/user.class.js
@@ -43,7 +43,6 @@ class User {
 
       // 게임 세션에서 유저 삭제
       userSession.removeUser(this.socket);
-      // 세션의 유저 배열이 빈 배열이면 세션과 인터벌도 삭제
 
       this.socket.destroy(); // 서버에서 소켓 종료
     }
