+++ src/handlers/index.js
@@ -3,9 +3,9 @@ import CustomError from '../utils/error/customError.js';
 import { ErrorCodes } from '../utils/error/errorCodes.js';
 import { animationHandler } from './social/playerAnimation.handler.js';
 import { chatHandler } from './social/playerChat.handler.js';
-import playerMoveHandler from './town/playerMove.handler.js';
-import townEnterHandler from './town/townEnter.handler.js';
-import playerLocationUpdateHandler from './town/playerLocationUpdate.handler.js';
+import playerMoveHandler from './transport/playerMove.handler.js';
+import townEnterHandler from './transport/townEnter.handler.js';
+import playerLocationUpdateHandler from './transport/playerLocationUpdate.handler.js';
 import registerHandler from './account/register.handler.js';
 import loginHandler from './account/login.handler.js';
 import createCharacterHandler from './account/createCharacter.handler.js';
@@ -19,7 +19,7 @@ import { disbandPartyHandler } from './social/party/disbandParty.handler.js';
 import { kickOutPartyHandler } from './social/party/kickOutParty.handler.js';
 import { allowInviteHandler } from './social/party/allowInvite.handler.js';
 import { leavePartyHandler } from './social/party/leaveParty.handler.js';
-import moveSectorHandler from './town/moveSectorHandler.js';
+import moveSectorHandler from './transport/moveSectorHandler.js';
 
 import { gatheringSkillCheckHandler } from './gathering/GatheringSkillCheck.handler.js';
 import { StartGatheringHandler } from './gathering/StartGathering.handler.js';
