+++ src/handlers/transport/playerMove.handler.js
@@ -1,12 +1,9 @@
-import { config } from '../../config/config.js';
-import { getGameAssets } from '../../init/assets.js';
 import { getNaveMesh } from '../../init/navMeshData.js';
 import { findPath } from '../../init/navMeshLoader.js';
-import { getSectorSessions, getPlayerSession } from '../../session/sessions.js';
+import { getPlayerSession } from '../../session/sessions.js';
 import CustomError from '../../utils/error/customError.js';
 import { ErrorCodes } from '../../utils/error/errorCodes.js';
 import handleError from '../../utils/error/errorHandler.js';
-import PACKET from '../../utils/packet/packet.js';
 
 // 클라이언트상에서 어떤 지점을 클릭했을 때 실행
 export async function playerMoveHandler(socket, packetData) {
@@ -48,10 +45,6 @@ export async function playerMoveHandler(socket, packetData) {
     } else {
       isValidPath = false;
     }
-
-    // Jmeter용 빈 응답 패킷
-    // const packet = PACKET.S2CPlayerMove();
-    // socket.write(packet);
 
     return isValidPath;
   } catch (error) {
